+++ ThunderStore/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.2.2
+- Change the throw distance of throwing the coin from 12 to 10
+- Changed the distance coins eject from the washing machine from 12 to 5 and added a config option to change the distance
+
 ## 1.2.1
 - Removed despawn script mechanic, washing machine will now just take scrap value off of items and give coins
 - Any item worth 0 scrap value will now no longer despawn from the ship on team wipe